Change input types from int to float for calculations

+++ 02.PYHON_IFMG/SEMANA01/s01_item_c.py
@@ -1,7 +1,7 @@
 print("----------CÁLCULO DE JUROS SIMPLES----------")
-c = int(input("Digite o valor de C:"))
-i = int(input("Digite o valor de I:"))
-t = int(input("Digite o valor de T:"))
+c = float(input("Digite o valor de C:"))
+i = float(input("Digite o valor de I:"))
+t = float(input("Digite o valor de T:"))
 
 J = (c*i*t)/100
 print(f"O valor de J é igual a:{J}")
